requirements.txt
opencv-python==4.11.0.86
pillow==11.2.1
face-recognition==1.3.0
dlib==19.24.8
cmake==4.0.2